+++ src/main/java/com/example/datn/entity/sale_on/CartItem.java
@@ -1,4 +1,4 @@
-package com.example.datn.entity;
+package com.example.datn.entity.sale_on;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,14 +7,13 @@ import lombok.Setter;
 
 import java.math.BigDecimal;
 
-@NoArgsConstructor
 @AllArgsConstructor
-@Getter
-@Setter
-public class CartItemRequest {
+@NoArgsConstructor
+@Getter@Setter
+public class CartItem {
     private Long productId;
     private String name;
-    private BigDecimal price;
+    private Double price;
     private int quantity;
     private BigDecimal discountedPrice;
 }
