--- src/main/java/com/example/datn/service/PhieuGiamGiaService.java
@@ -1,5 +1,0 @@
-package com.example.datn.service;
-
-public class PhieuGiamGiaService {
-
-}
